chore: move fee related function to fee.go
