+++ x/exchange/types/fee_test.go
@@ -1,11 +1,15 @@
 package types_test
 
 import (
+	"fmt"
+	"math/rand"
 	"testing"
 
-	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/stretchr/testify/require"
 
+	sdk "github.com/cosmos/cosmos-sdk/types"
+
+	"github.com/crescent-network/crescent/v5/app/testutil"
 	utils "github.com/crescent-network/crescent/v5/types"
 	"github.com/crescent-network/crescent/v5/x/exchange/types"
 )
@@ -67,5 +71,34 @@ func TestValidateFeeRates(t *testing.T) {
 				require.EqualError(t, err, tc.expectedErr)
 			}
 		})
+	}
+}
+
+func TestDeductFee(t *testing.T) {
+	for i, tc := range []struct {
+		amt, feeRate  sdk.Dec
+		deducted, fee sdk.Dec
+	}{
+		{
+			utils.ParseDec("123456789"), utils.ParseDec("0.003"),
+			utils.ParseDec("123086418.633"), utils.ParseDec("370370.367"),
+		},
+		{
+			utils.ParseDec("123456789"), utils.ParseDec("0.0015"),
+			utils.ParseDec("123271603.8165"), utils.ParseDec("185185.1835"),
+		},
+	} {
+		t.Run(fmt.Sprint(i), func(t *testing.T) {
+			deducted, fee := types.DeductFee(tc.amt, tc.feeRate)
+			testutil.AssertEqual(t, tc.deducted, deducted)
+			testutil.AssertEqual(t, tc.fee, fee)
+		})
+	}
+
+	r := rand.New(rand.NewSource(1))
+	for i := 0; i < 50; i++ {
+		amt := utils.RandomDec(r, sdk.NewDec(10), sdk.NewDec(100000000))
+		deducted, fee := types.DeductFee(amt, utils.ParseDec("0.003"))
+		testutil.AssertEqual(t, amt, deducted.Add(fee))
 	}
 }
